Change decodeIp to only consider use up to byte Length bytes.
This solves the issue where Ethernet packets must be a minimum of 60 bytes, but
can take up only 54 bytes on the wire if there's no payload added.  Without
this change, packets of that type may appear to be TCP packets with 6 null bytes
in their payload.

See
http://forums.devshed.com/networking-help-109/tcp-protocol-mysterious-6-null-byte-payload-303357.html
for more discussion on this.