[RFC] Add reassembly package and accompanying reassemblydump example
This is the current state of our reassembly package which does full
(IPv4 and TCP) bidirectional reassembly, implements a TCP state machine
and various security checks on TCP traffic. It is based on the existing
tcpassembly package. A reassemblydump program is provided which uses the
reassembly package, with support for DNS and HTTP reassembly.

This is the second submission of this code after a first round of review
by Graeme. It is probably still not fit for merging as-is, however all
comments and feedback are welcome at this stage.