Stop pointers-of-pointers from escaping with cgo.
Yet more evidence that escape analysis in Go is really weird.  It
appears that the pointers to pkthdr and bufptr were being allocated on
the heap... attaching them to the handle itself appears to have helped
out with this.

This has pretty hacky cast-to-uintptr magic in it to get around cgo's
error:

  panic: runtime error: cgo argument has Go pointer to Go pointer