reassembly: Fix offset calculation in reassemblyObject.CaptureInfo
CaptureInfo is supposed to return info of packet corresponding to the
given offset. The current implementation does not take the livePacket
or page length into account, so any offset other than the first offset
of a packet will get the CaptureInfo of the next packet.

This PR also removes a redundant `CaptureInfo` field within
`livePacket`.